+++ lightsim2grid/tests/test_NICSLUSolver.py
@@ -4,10 +4,11 @@ import numpy as np
 import pdb
 import zipfile
 from scipy import sparse
-KLU_AVAILBLE = False
+import re
+NICSLU_AVAILBLE = False
 try:
-    from lightsim2grid_cpp import KLUSolver
-    KLU_AVAILBLE = True
+    from lightsim2grid_cpp import NICSLUSolver
+    NICSLU_AVAILBLE = True
 except ImportError:
     # KLU solver is not available, these tests cannot be carried out
     pass
@@ -21,9 +22,9 @@ class MakeTests(unittest.TestCase):
         self.max_it = 10
         self.tol = 1e-8  # tolerance for the solver
         self.tol_test = 1e-4  # tolerance for the test (2 matrices are equal if the l_1 of their difference is less than this)
-        if not KLU_AVAILBLE:
+        if not NICSLU_AVAILBLE:
             return
-        self.solver = KLUSolver()
+        self.solver = NICSLUSolver()
 
         self.path = None
         self.V_init = None
@@ -49,7 +50,7 @@ class MakeTests(unittest.TestCase):
                 self.Ybus = self.load_array(myzip, "Ybus.npy")
                 self.Ybus = sparse.csc_matrix(self.Ybus)
             return True
-        except:
+        except Exception as exc_:
             return False
 
     def _load_vect(self, myzip, iter_max, nm):
@@ -98,8 +99,8 @@ class MakeTests(unittest.TestCase):
         assert np.sum(np.abs(Vm - Vm_pp)) <= self.tol_test, "voltages magnitude are not the same"
 
     def test_dir(self):
-        if not KLU_AVAILBLE:
-            self.skipTest("KLU is not installed")
+        if not NICSLU_AVAILBLE:
+            self.skipTest("NICSLU is not installed")
         nb_tested = 0
         for path in os.listdir("."):
             _, ext = os.path.splitext(path)